requirements.txt
flask
flask-sqlalchemy
mysql-connector-python
google-api-python-client
openai
gspread
oauth2client